refactor: reorganized search components, moved base UI components, import paths updated
