--- src/app/components/Filter/index.js
@@ -1,2 +1,0 @@
-export * from "./Filter";
-export { default } from "./Filter";
