--- src/app/components/ResultsList/index.js
@@ -1,2 +1,0 @@
-export * from "./ResultsList";
-export { default } from "./ResultsList";
