+++ src/app/components/Search/SearchFilter/FilterContainer/index.js
@@ -1,0 +1,2 @@
+export * from "./FilterContainer";
+export { default } from "./FilterContainer";
