+++ src/app/components/Search/SearchResults/ResultsCard/index.js
@@ -1,0 +1,2 @@
+export * from "./ResultsCard";
+export { default } from "./ResultsCard";
