--- src/app/components/SearchBar/index.js
@@ -1,2 +1,0 @@
-export * from "./SearchBar";
-export { default } from "./SearchBar";
