--- src/app/components/SearchResults/index.js
@@ -1,2 +1,0 @@
-export * from "./SearchResults";
-export { default } from "./SearchResults";
