+++ src/app/components/UI/SegmentedControl/index.js
